requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit>=1.28.0
joblib>=1.3.0
plotly>=5.17.0
requests>=2.31.0
python-dotenv>=1.0.0
pyarrow>=14.0.0
schedule>=1.2.0
pytz>=2024.1
earthengine-api>=0.1.380
shap>=0.44.0
